refactor: modified github worker to add the deserializer inside the worker
